Embed in SpeziHealthKit Test Suite
